"gradient_checkpointing": False

+++ train.py
@@ -95,7 +95,7 @@ if __name__ == '__main__':
     # 加载sft权重
     # pl_model.load_sft_weight('./best_ckpt/best.pt',is_trainable=True)
 
-    # pl_model = pl_model.float() if not is_bf16_supported else pl_model.bfloat16()
+    pl_model = pl_model.float()
 
     def dataset_loader_filter_fn(dataset):
         print('*' * 30, 'total', len(dataset))
